requirements.txt
astroid==2.4.2
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
cycler==0.10.0
idna==2.10
isort==4.3.21
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
matplotlib==3.3.3
mccabe==0.6.1
numpy==1.19.1
pandas==1.1.0
Pillow==8.1.0
pylint==2.5.3
pyotp==2.4.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-decouple==3.4
python-dotenv==0.15.0
pytz==2020.1
requests==2.24.0
robin-stocks==1.6.3
six==1.15.0
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.10
wrapt==1.12.1
